Extract method-specific routines to separate classes and functions
